增加根据房间id获取有序试题列表接口

+++ course-web-server/src/main/java/com/huatu/tiku/course/service/v1/practice/CoursePracticeQuestionInfoService.java
@@ -20,5 +20,5 @@ public interface CoursePracticeQuestionInfoService extends BaseServiceHelper<Cou
 	 * @param roomId
 	 * @return
 	 */
-	List<Long> getQuestionsInfoByRoomId(Long roomId);
+	List<Integer> getQuestionsInfoByRoomId(Long roomId);
 }
